544 footnote support working on iOS
This commit includes the newsfoot.js script in the iOS page.html file
and copies css from the Mac styleSheet.css into the iOS styleSheet.css.

The is only one difference in the styles between the two: iOS defines left and
right to be in terms of the viewport width rather than a fixed size. This ensures that the popover is always within the visible area.