Improves compatibility with sharing services
In relation to #3123: the changes in this commit allow text to be shared to additional apps in the shae menu.

The first change is the ordering of the `writableTypes` array. It now puts `.string` first. This change on its own allows text to be shared to Tweetbot, Ulysees, and potentially other apps.

The second change is to the text—`plainText()`—that is shared. It has been reduced to `title` and `url`, which makes more sense in the context of sharing snippets. This is in large part to do with the fact that sharing title, contentText, summary, contentHTML, url, externalURL, date, and feedName is somewhat overkill when sharing to a Twitter-like service.