Add http and https to the list of CFBundleURLSchemes.
If NetNewsWire says it is able to handle http and https URL schemes, Safari will list NetNewsWire as a target in Develop > Open Page With, which provides a trivial flow for subscribing to the page or site one is currently reading in Safari. I believe this functionality existed in NetNewsWire "classic," and this patch just brings the functionality back.