Add http and https to the list of CFBundleURLSchemes.
If NetNewsWire says it is able to handle http and https URL schemes, Safari will list NetNewsWire as a target in Develop > Open Page With, which provides a trivial flow for subscribing to the page or site one is currently reading in Safari. This has the advantage of providing a reasonably fast path for subscription without requiring users configure the app extension to get the functionality.